实现 sendfile 系统调用, 修复部分警告. 关闭函数形参的 unused 警告 (大量 syscall 宏被误报) 诞生无警告文件 id_alloc.c 和 ipc.c.

+++ src/util/id_alloc.c
@@ -33,7 +33,7 @@ int32_t id_alloc(id_allocator_t *allocator) {
             allocator->bitmap[word_index] |= mask;
             allocator->free_count--;
             allocator->next_id = (id + 1) % allocator->max_ids;
-            return id;
+            return (int32_t)id;
         }
 
         id = (id + 1) % allocator->max_ids;
